Removed StateMachine management of the RenderWindow
Before this update, the StateMachine had a reference to the
RenderWindow. and controlled clearing and displaying of the renderer, as
well as the execution order via StateMachine.Execute(). This is
significantly more responsibility than necessary for a StateMachine. It
also controlled initial input processing, and a UserQuit variable. It
now no longer does any of this.

Summary:

- The StateMachine.Execute() function has been removed, and
StateMachine.UpdateStates() and StateMachine.DrawStates() have been made
public
- The StateMachine no longer carries a reference to an sf::RenderWindow,
and thus no longer requires one to be passed via the constructor.
- StateMachine.DrawStates() now requires the window to be passed, as it
no longer carries the reference itself.
- StateMachine.ProcessInput() removed
- StateMachine.GetUserQuit() and _userQuit removed.
- Demo code contains sample input processing and game loop with
replacement 'quit' variable
- Demo code, documentation and README.md have been updated accordingly.